mejoras

+++ src/main/java/com/factura/FacturaElectronica/Bo/ArticuloBo.java
@@ -1,5 +1,7 @@
 package com.factura.FacturaElectronica.Bo;
 
+import java.math.BigDecimal;
+
 import com.factura.FacturaElectronica.modelo.Articulo;
 import com.factura.FacturaElectronica.modelo.Empresa;
 
@@ -17,6 +19,6 @@ public interface ArticuloBo {
 
 	Articulo buscarPorCodigoYEmpresa(String codigo, Empresa empresa);
 
-	Double porcentanjeDeGanancia(Double costo, Double iva, Double precio);
+	BigDecimal porcentanjeDeGanancia(BigDecimal costo, BigDecimal iva, BigDecimal precio);
 
 }
