+++ src/main/java/com/factura/FacturaElectronica/Dao/InventarioDao.java
@@ -1,5 +1,7 @@
 package com.factura.FacturaElectronica.Dao;
 
+import java.math.BigDecimal;
+
 import com.factura.FacturaElectronica.modelo.Articulo;
 import com.factura.FacturaElectronica.modelo.Inventario;
 
@@ -15,7 +17,7 @@ public interface InventarioDao {
 
 	Inventario buscarPorArticulo(Articulo articulo);
 	Inventario findByArticuloAndEstado(Articulo articulo,String estado);
-	Double getTotalCosto(Inventario inventario, Double cantidad);
+	BigDecimal getTotalCosto(Inventario inventario, BigDecimal cantidad);
