+++ src/main/java/com/factura/FacturaElectronica/Dao/KardexDao.java
@@ -1,5 +1,7 @@
 package com.factura.FacturaElectronica.Dao;
 
+import java.math.BigDecimal;
+
 import com.factura.FacturaElectronica.modelo.Inventario;
 import com.factura.FacturaElectronica.modelo.Kardex;
 import com.factura.FacturaElectronica.modelo.Usuario;
@@ -8,8 +10,8 @@ public interface KardexDao {
 
 	void agregar(Kardex Kardex);
 	
-	void entrada(Inventario inventario, Double cantidad, String observacion, String consecutivo, String tipo, String motivo, Usuario usuarioSesion);
-	void salida(Inventario inventario, Double cantidad, String observacion, String consecutivo, String tipo, String motivo,Usuario usuario);
+	void entrada(Inventario inventario, BigDecimal cantidad, String observacion, String consecutivo, String tipo, String motivo, Usuario usuarioSesion);
+	void salida(Inventario inventario, BigDecimal cantidad, String observacion, String consecutivo, String tipo, String motivo,Usuario usuario);
