+++ src/main/java/com/factura/FacturaElectronica/Dao/UsuarioCajaDao.java
@@ -1,5 +1,7 @@
 package com.factura.FacturaElectronica.Dao;
 
+import java.math.BigDecimal;
+
 import com.factura.FacturaElectronica.modelo.Usuario;
 import com.factura.FacturaElectronica.modelo.UsuarioCaja;
 
@@ -15,6 +17,6 @@ public interface UsuarioCajaDao {
 	
 	UsuarioCaja findByUsuarioAndEstado(Usuario usuario,String estado);
 	
-	void actualizarCaja(UsuarioCaja usuarioCaja ,Double totalEfectivo,Double totalTarjeta,Double totalBanco,Double totalCredito,Double totalAbono);
+	void actualizarCaja(UsuarioCaja usuarioCaja ,BigDecimal totalEfectivo,BigDecimal totalTarjeta,BigDecimal totalBanco,BigDecimal totalCredito,BigDecimal totalAbono);
 
 }
